# [GC2023] Simple fix for allocation overhead