--- a/bench_test.go
+++ b/bench_test.go
@@ -1,18 +1,37 @@
 package main
 
-import "testing"
+import (
+	_ "embed"
+	"testing"
+)
+
+//go:embed caisson.go
+var trigramexample string
 
 func BenchmarkTrigrams(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		_ = Trigrams("this is a test")
+		_ = Trigrams(trigramexample)
 	}
 }
 
-func BenchmarkTokenize(b *testing.B) {
-	const text = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.`
-	var result []Trigram
-	for n := 0; n < b.N; n++ {
-		result = Tokenize(text)
+func BenchmarkTrigramBytes(b *testing.B) {
+	trigrams := Trigrams(trigramexample)
+
+	for i := 0; i < b.N; i++ {
+		trigrams[i%len(trigrams)].Bytes()
 	}
-	_ = result
+}
+
+func BenchmarkTrigramBytesFast(b *testing.B) {
+	trigrams := Trigrams(trigramexample)
+
+	for i := 0; i < b.N; i++ {
+		trigrams[i%len(trigrams)].BytesFast()
+	}
+}
+
+func BenchmarkItemise(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		_ = Itemise(Tokenize(trigramexample))
+	}
 }