A type, `Trigram` is defined as 3 runes. This saves time spent in the expensive allocation. Benchmarks show a **2.6x** improvement with this simple fix alone.

**Before:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10             3804986               310.9 ns/op
BenchmarkTokenize-10              104250             11147 ns/op
PASS
ok      indexer 2.885s
```

**After:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10            22071332                48.99 ns/op
BenchmarkTokenize-10              273313              4267 ns/op
PASS
ok      indexer 2.438s
```