@boyter 2023-11-12
Sweet. Im going to do a benchmark of all of the PR's against my best effort and turn it into a blog post, and then merge the winner.