@jamesrom 2023-11-13
Nice, I look forward to reading the post. This fix is simple enough that it should work with the other fixes that have been proposed too.

By the way, I ran the benchmarks again with `-benchmem` and you can see a **13x** improvement on allocations in `Trigrams`:


**Before:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10             3684786               311.7 ns/op           288 B/op         13 allocs/op
BenchmarkTokenize-10              105349             11190 ns/op           21608 B/op        312 allocs/op
PASS
ok      indexer 3.181s
```

**After:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10            24047827                49.17 ns/op          144 B/op          1 allocs/op
BenchmarkTokenize-10              278169              4193 ns/op           12904 B/op         72 allocs/op
PASS
ok      indexer 3.872s
```