@boyter 2023-11-13
The only catch I see is that you arent calling .Bytes on the function in your benchmark, which of course means the output is different.

Was your intention to just have that evaluated later in the Itemise function when it is hashing?

For the moment I added a quick change to master where it includes a copy of this so you can compare this against a few other methods,

```
hyperfine 'indexer --trigram=default' 'indexer --trigram=jamesrom'
```

As you can see its not any faster at the moment because its calling the .Bytes inside it.