@jamesrom 2023-11-14
Yeah, the main idea is that you can pass around Trigrams not strings, as you don't need to worry about string encoding if you're just hashing the bytes. Then you can do stuff like what I've done in bc2ecbf13b9e88ce1bd703d2afa6bd4d3f357a1e (BytesFast is 25x faster than Bytes).

I've added more complete test `BenchmarkItemise` to show how this works.

**Before:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10                6596            180396 ns/op          226217 B/op       7799 allocs/op
BenchmarkItemise-10                 2503            466337 ns/op          617474 B/op      15979 allocs/op
PASS
ok      indexer 3.797s
```

**After:**
```
goos: darwin
goarch: arm64
pkg: indexer
BenchmarkTrigrams-10                       49664             23261 ns/op          131072 B/op          2 allocs/op
BenchmarkTrigramBytes-10                74693443                15.84 ns/op            0 B/op          0 allocs/op
BenchmarkTrigramBytesFast-10            1000000000               0.6242 ns/op          0 B/op          0 allocs/op
BenchmarkItemise-10                         3018            383156 ns/op          510642 B/op      12348 allocs/op
PASS
ok      indexer 4.896s
```

**1.22x** improvement in ns/op
**1.29x** improvement in allocs/op